Flowseal - Handmade Bot/requirements.txt
aiohttp==3.6.3
PyNaCl==1.3.0
async-timeout==3.0.1
asyncio==3.4.3
beautifulsoup4==4.6.0
bs4==0.0.1
discord.py==1.5.1
docopt==0.6.2
requests==2.23.0
Telethon==1.17.5
urllib3==1.24.2
youtube-dl==2021.3.3
ffmpeg-python==0.2.0
PyGithub==1.54.1